perf: cut generation time ~25% and allocation ~66% in the matrix pipeline
- store MatrixData modules as byte[][] instead of int[][] (breaking:
  getMatrix() now returns byte[][])
- reuse a single scratch matrix across the 8 mask trials instead of
  copying the base matrix per trial; reapply the winning mask at the end
- add a penalty cutoff to PenaltyCalculator so trials abort as soon as
  they are provably worse than the current best mask
- scan penalty columns in place instead of materializing a transposed
  copy per trial
- precompute the zig-zag placement order once per version, cached in the
  generator pipeline, instead of re-walking it on every mask trial
- encode the payload to UTF-8 once per generate() call
- fix getAsImage: use MemoryCacheImageOutputStream (no temp file per
  call), fail loudly when no ImageIO writer handles the format, and
  throw UncheckedIOException instead of RuntimeException

+++ src/main/java/com/qrlib/matrix/FormatInformation.java
@@ -27,7 +27,7 @@ final class FormatInformation {
     }
 
     static void write(MatrixData matrixData, ECCLevel eccLevel, int maskPattern) {
-        int[][] matrix = matrixData.getMatrix();
+        byte[][] matrix = matrixData.getMatrix();
         int size = matrix.length;
 
         int[] formatBits = computeFormatBits(eccLevel, maskPattern);
@@ -40,12 +40,12 @@ final class FormatInformation {
 
         // Mirrored sequence across the top-right and bottom-left finder patterns
         for (int i = 0; i < FORMAT_INFO_BITS; i++) {
-            matrix[pos1[i][0]][pos1[i][1]] = formatBits[i];
+            matrix[pos1[i][0]][pos1[i][1]] = (byte) formatBits[i];
 
             if (i < QrLayout.FINDER_WITH_SEPARATOR_SIZE) {
-                matrix[QrLayout.FINDER_WITH_SEPARATOR_SIZE][size - 1 - i] = formatBits[i];
+                matrix[QrLayout.FINDER_WITH_SEPARATOR_SIZE][size - 1 - i] = (byte) formatBits[i];
             } else {
-                matrix[size - FORMAT_INFO_BITS + i][QrLayout.FINDER_WITH_SEPARATOR_SIZE] = formatBits[i];
+                matrix[size - FORMAT_INFO_BITS + i][QrLayout.FINDER_WITH_SEPARATOR_SIZE] = (byte) formatBits[i];
             }
         }
     }
