+++ src/test/java/com/qrlib/CharacterizationTest.java
@@ -52,13 +52,13 @@ class CharacterizationTest {
                 .version(version)
                 .ECCLevel(eccLevel)
                 .build();
-        int[][] matrix = generator.generate(PAYLOAD).getMatrixData().getMatrix();
+        byte[][] matrix = generator.generate(PAYLOAD).getMatrixData().getMatrix();
         return matrixHash(matrix);
     }
 
-    private static String matrixHash(int[][] matrix) {
+    private static String matrixHash(byte[][] matrix) {
         StringBuilder sb = new StringBuilder();
-        for (int[] row : matrix) {
+        for (byte[] row : matrix) {
             for (int cell : row) {
                 sb.append(cell);
             }
